Login funcional + javadoc Server simulat localment
