+++ src/main/java/fithub/clientEscriptori/Main.java
@@ -2,6 +2,13 @@ package fithub.clientEscriptori;
 
 import fithub.clientEscriptori.app.ControladorAplicacio;
 
+/**
+ * Classe Main punt d'entrada de l'aplicació
+ *
+ *
+ * @author Xavi Cano Garcia
+ * @version 1.0
+ */
 public class Main {
     public static void main(String[] args) {
         ControladorAplicacio controladorAplicacio = new ControladorAplicacio();
