+++ src/main/java/fithub/clientEscriptori/gui/ControladorGui.java
@@ -1,11 +1,20 @@
 package fithub.clientEscriptori.gui;
-
+/**
+ * Clase arrel de l'interficie gràfica.
+ * Conté tots els elements grafics
+ * Executa accion com canviar de pantalla
+ *
+ * @author Xavi Cano Garcia
+ * @version 1.0
+ */
 public class ControladorGui {
     LoginFrame loginFrame;
     MainFrame mainFrame;
 
     ControladorPanells controladorPanells;
-
+    /**
+     * Constructor objecte ControladorGui.
+     */
     public ControladorGui() {
         controladorPanells = new ControladorPanells();
 
@@ -14,7 +23,12 @@ public class ControladorGui {
         loginFrame.add(controladorPanells.loginForm.getPanel1());
 
     }
-
+    /**
+     * Metode que executa un canvi de pantalla.
+     *
+     * @param frame Frame al que es vol canviar.
+     * @param userType Tipus de usuari que es fara servir per triar el panell.
+     */
     public void canviaPantalla(String frame, String userType) {
         if (frame.equals("main")) {
             loginFrame.setVisible(false);
