+++ src/main/java/fithub/clientEscriptori/gui/ControladorPanells.java
@@ -3,14 +3,21 @@ package fithub.clientEscriptori.gui;
 import fithub.clientEscriptori.gui.panells.LoginForm;
 import fithub.clientEscriptori.gui.panells.MainAdmin;
 import fithub.clientEscriptori.gui.panells.MainUser;
-
+/**
+ * Clase que conte y selecciona tots els panells.(inprogress)
+ *
+ * @author Xavi Cano Garcia
+ * @version 1.0
+ */
 public class ControladorPanells {
     LoginForm loginForm;
 
     MainUser mainUser;
 
     MainAdmin mainAdmin;
-
+    /**
+     * Constructor objecte Controlador de panells.
+     */
     public ControladorPanells() {
 
         loginForm = new LoginForm();
