+++ src/main/java/fithub/clientEscriptori/gui/MainFrame.java
@@ -1,11 +1,18 @@
 package fithub.clientEscriptori.gui;
 
 import javax.swing.*;
-
+/**
+ * Clase que defineix la finestra principal.
+ *
+ * @author Xavi Cano Garcia
+ * @version 1.0
+ */
 public class MainFrame extends JFrame {
 
     JPanel currentPanel;
-
+    /**
+     * Constructor objecte Main frame.
+     */
     public MainFrame() {
         this.setTitle("FITHUB - Main");
         this.setSize(800, 600);
@@ -17,7 +24,11 @@ public class MainFrame extends JFrame {
         add(currentPanel);
 
     }
-
+    /**
+     * Metode que canvia el panell del Main framme.
+     *
+     * @param nouPanell Panell nou que es vol canviar.
+     */
     public void canviPanell(JPanel nouPanell) {
 
         remove(currentPanel);
